@@ -4,9 +4,7 @@
   "private": true,
   "scripts": {
     "server": "npx ts-node server/server.js",
-    "test": "react-app-rewired test --watchAll",
-    "coverage": "npm test -- --verbose --coverage --watchAll",
-    "test-clear": "npm test -- --clearCache",
+    "lint:fix": "eslint --fix --ext .js,.ts",
     "keys": "node -e 'console.log(`key: ${crypto.randomBytes(32).toString(\"hex\")}\n IV: ${crypto.randomBytes(16).toString(\"hex\")}`)'"
   },
   "dependencies": {
@@ -23,8 +21,7 @@
     "express-rate-limit": "^7.0.2",
     "js-logger": "^1.6.1",
     "redis": "^4.6.12",
-    "stream-browserify": "^3.0.0",
-    "typescript": "^5.0.4"
+    "stream-browserify": "^3.0.0"
   },
   "eslintConfig": {
     "extends": [
@@ -51,13 +48,20 @@
     "@babel/preset-env": "^7.23.8",
     "@babel/preset-react": "^7.18.6",
     "@types/express": "^4.17.21",
-    "ts-node": "^10.9.2",
+    "@typescript-eslint/eslint-plugin": "^6.20.0",
     "babel-plugin-js-logger": "^1.0.17",
     "customize-cra": "^1.0.0",
     "cypress": "^13.6.2",
+    "eslint": "^8.56.0",
     "eslint-config-mx": "^4.0.0",
+    "eslint-config-standard-with-typescript": "^43.0.1",
     "eslint-plugin-cypress": "^2.15.1",
-    "eslint-plugin-redux-namespaced-actions": "^1.0.0",
-    "ts-jest": "^29.1.2"
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^27.6.3",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.1.1",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   }
 }
