@@ -1,22 +1,22 @@
-import { get } from 'axios';
+import { get } from 'axios'
 
-const baseUrl = 'http://127.0.0.1:8080/';
+const baseUrl = 'http://127.0.0.1:8080/'
 
 describe('Test Routes', () => {
   test('Default route', async () => {
-    const res = await get(baseUrl);
+    const res = await get(baseUrl)
 
-    expect(res).toBeTruthy();
-    expect(res.status).toBe(200);
-    expect(res.data).toContain('<!DOCTYPE html>');
-    expect(res.data).toContain('Web site created using create-react-app');
-  });
+    expect(res).toBeTruthy()
+    expect(res.status).toBe(200)
+    expect(res.data).toContain('<!DOCTYPE html>')
+    expect(res.data).toContain('Web site created using create-react-app')
+  })
 
   test('PING route', async () => {
-    const res = await get(`${baseUrl}ping`);
+    const res = await get(`${baseUrl}ping`)
 
-    expect(res).toBeTruthy();
-    expect(res.status).toBe(200);
-    expect(res.data).toEqual('ok');
-  });
-});
+    expect(res).toBeTruthy()
+    expect(res.status).toBe(200)
+    expect(res.data).toEqual('ok')
+  })
+})
