@@ -1,9 +1,9 @@
-const processEnv = {};
-const envs = {...process.env, ...process.client_envs};
+const processEnv = {}
+const envs = { ...process.env, ...process.client_envs }
 
 Object.keys(envs).forEach((k) => {
-  processEnv[k.toUpperCase()] = envs[k];
-});
+  processEnv[k.toUpperCase()] = envs[k]
+})
 
 const config = {
   AuthServiceEndpoint: 'https://login.universalconnectproject.org/api',
@@ -20,8 +20,8 @@ const config = {
   Env: 'pre', // mocked
   Version: '',
   UseAxios: true,
-  CryptoKey: '', //crypto.randomBytes(32).toString("hex");
-  CryptoIv: '', //crypto.randomBytes(16).toString("hex");
+  CryptoKey: '', // crypto.randomBytes(32).toString("hex");
+  CryptoIv: '', // crypto.randomBytes(16).toString("hex");
   CryptoAlgorithm: 'aes-256-cbc',
   RedisServer: 'redis://localhost:6379',
   RedisCacheTimeSeconds: 600,
@@ -29,19 +29,19 @@ const config = {
   // ResourcePrefix: 'local',
   ResourceVersion: '', // 'development'
 
-  SophtronClientId: '', //sophtron.com -> api settings -> direct auth -> UserId
-  SophtronClientSecret: '',  //sophtron.com -> api settings -> direct auth -> AccessKey
+  SophtronClientId: '', // sophtron.com -> api settings -> direct auth -> UserId
+  SophtronClientSecret: '', // sophtron.com -> api settings -> direct auth -> AccessKey
 
   UcpAuthClientId: '', // https://login.universalconnectproject.org/
   UcpAuthClientSecret: '',
   UcpAuthEncryptionKey: ''
 
-};
-
-const arr = Object.keys(config);
-for (let i = 0; i < arr.length; i++) {
-  const key = arr[i];
-  config[key] = processEnv[key.toUpperCase()] || config[key];
 }
 
-module.exports = config;
+const arr = Object.keys(config)
+for (let i = 0; i < arr.length; i++) {
+  const key = arr[i]
+  config[key] = processEnv[key.toUpperCase()] || config[key]
+}
+
+module.exports = config
