@@ -1,10 +1,10 @@
 describe('Test Connect Routes', () => {
   test('Test sendAnalyticsEvent route', async () => {
     // TODO: Figure out error with mock
-    expect(true).toBe(true);
+    expect(true).toBe(true)
     // const res = mocked.get('sendAnalyticsEvent', () => 'ok');
     //
     // expect(res).toBeTruthy();
     // expect(res.status).toBe(200);
-  });
-});
+  })
+})
