@@ -1,7 +1,7 @@
 const fs = require('fs')
-const {ApiEndpoints} = require('../../shared/connect/ApiEndpoint.js')
+const { ApiEndpoints } = require('../../shared/connect/ApiEndpoint.js')
 
-module.exports = function(app){
+module.exports = function (app) {
   app.get('/jobs/:guid', async (req, res) => {
     res.sendFile(__dirname + '/stubs/job.json')
   })
@@ -21,7 +21,7 @@ module.exports = function(app){
     res.sendFile(__dirname + '/stubs/member.json')
   })
   app.get(`${ApiEndpoints.MEMBERS}/:member_guid/credentials`, async (req, res) => {
-    let ret = await connectService.getMemberCredentials(req.params.member_guid)
+    const ret = await connectService.getMemberCredentials(req.params.member_guid)
     res.send(ret)
   })
   app.get(`${ApiEndpoints.INSTITUTIONS}/:institution_guid/credentials`, async (req, res) => {
