@@ -1,15 +1,15 @@
 const fs = require('fs')
-const {ApiEndpoints} = require('../../shared/connect/ApiEndpoint.js')
+const { ApiEndpoints } = require('../../shared/connect/ApiEndpoint.js')
 
-module.exports = function(app){
+module.exports = function (app) {
   app.post('/feature_visits', async (req, res) => {
     res.sendStatus(200)
   })
   app.get(ApiEndpoints.AGREEMENT, async (req, res) => {
-    res.send("")
+    res.send('')
   })
   app.get('/offers/*', async (req, res) => {
-    res.send("")
+    res.send('')
   })
   app.get(ApiEndpoints.USER_FEATURES, async (req, res) => {
     res.send(require('./stubs/user_features.js'))
