@@ -1,10 +1,10 @@
 module.exports = {
-  "analytics_session": {
-    "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-    "is_first_visit": false,
-    "browser_name": "Chrome",
-    "browser_version": "111.0.0.0",
-    "product_name": "Individual Widget",
-    "product_version": "1.0.0"
+  analytics_session: {
+    guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+    is_first_visit: false,
+    browser_name: 'Chrome',
+    browser_version: '111.0.0.0',
+    product_name: 'Individual Widget',
+    product_version: '1.0.0'
   }
 }
