@@ -1,114 +1,114 @@
 module.exports = {
-  "member": {
-    "aggregation_status": 2,
-    "connection_status": 3,
-    "guid": "MBR-241266ed-803a-4841-8a8a-0f37551e8f56",
-    "institution_guid": "INS-241266ed-803a-4841-8a8a-0f37551e8f56",
-    "institution_name": null,
-    "institution_url": "https://gringotts.sand.internal.mx",
-    "instructional_data": {
-      "title": "Test Instructions",
-      "description": "<a href=\"https://google.com\" id=\"instructional_text\">A Test Link</a>",
-      "steps": [
-        "Go to <a href=\"https://google.com\" id=\"instructional_text_steps\">Google</a>"
+  member: {
+    aggregation_status: 2,
+    connection_status: 3,
+    guid: 'MBR-241266ed-803a-4841-8a8a-0f37551e8f56',
+    institution_guid: 'INS-241266ed-803a-4841-8a8a-0f37551e8f56',
+    institution_name: null,
+    institution_url: 'https://gringotts.sand.internal.mx',
+    instructional_data: {
+      title: 'Test Instructions',
+      description: '<a href="https://google.com" id="instructional_text">A Test Link</a>',
+      steps: [
+        'Go to <a href="https://google.com" id="instructional_text_steps">Google</a>'
       ]
     },
-    "is_being_aggregated": true,
-    "is_manual": false,
-    "is_managed_by_user": true,
-    "is_oauth": false,
-    "last_job_guid": "JOB-241266ed-803a-4841-8a8a-0f37551e8f56",
-    "last_job_status": null,
-    "last_update_time": null,
-    "metadata": null,
-    "mfa": {
-      "credentials": [
+    is_being_aggregated: true,
+    is_manual: false,
+    is_managed_by_user: true,
+    is_oauth: false,
+    last_job_guid: 'JOB-241266ed-803a-4841-8a8a-0f37551e8f56',
+    last_job_status: null,
+    last_update_time: null,
+    metadata: null,
+    mfa: {
+      credentials: [
         {
-          "nullify": null,
-          "guid": "CRD-241266ed-803a-4841-8a8a-0f37551e8f56",
-          "institution_guid": "INS-241266ed-803a-4841-8a8a-0f37551e8f56",
-          "label": "What city were you born in?",
-          "field_name": null,
-          "opt": null,
-          "type": 0,
-          "display_order": null,
-          "external_id": "UNIQUE_ID_FOR_THIS_CHALLENGE-f067cb4ab7",
-          "data_source_guid": null,
-          "field_type": 0,
-          "optional": null,
-          "editable": null,
-          "mfa": true,
-          "optional_mfa": null,
-          "aggregator_guid": "AGG-241266ed-803a-4841-8a8a-0f37551e8f56",
-          "question_field_type": null,
-          "answer_field_type": null,
-          "escaped": null,
-          "value_identifier": null,
-          "value_mask": null,
-          "size": null,
-          "max_length": null,
-          "created_at": "2023-03-30T22:45:43.000+00:00",
-          "updated_at": "2023-03-30T22:45:43.000+00:00",
-          "legacy_question_field_type": null,
-          "response_field_type": null,
-          "sequence": null,
-          "meta_data": null,
-          "display_name": null,
-          "status_code": 200,
-          "job_guid": "JOB-241266ed-803a-4841-8a8a-0f37551e8f56",
-          "options": []
+          nullify: null,
+          guid: 'CRD-241266ed-803a-4841-8a8a-0f37551e8f56',
+          institution_guid: 'INS-241266ed-803a-4841-8a8a-0f37551e8f56',
+          label: 'What city were you born in?',
+          field_name: null,
+          opt: null,
+          type: 0,
+          display_order: null,
+          external_id: 'UNIQUE_ID_FOR_THIS_CHALLENGE-f067cb4ab7',
+          data_source_guid: null,
+          field_type: 0,
+          optional: null,
+          editable: null,
+          mfa: true,
+          optional_mfa: null,
+          aggregator_guid: 'AGG-241266ed-803a-4841-8a8a-0f37551e8f56',
+          question_field_type: null,
+          answer_field_type: null,
+          escaped: null,
+          value_identifier: null,
+          value_mask: null,
+          size: null,
+          max_length: null,
+          created_at: '2023-03-30T22:45:43.000+00:00',
+          updated_at: '2023-03-30T22:45:43.000+00:00',
+          legacy_question_field_type: null,
+          response_field_type: null,
+          sequence: null,
+          meta_data: null,
+          display_name: null,
+          status_code: 200,
+          job_guid: 'JOB-241266ed-803a-4841-8a8a-0f37551e8f56',
+          options: []
         }
       ]
     },
-    "most_recent_job_guid": "JOB-241266ed-803a-4841-8a8a-0f37551e8f56",
-    "needs_updated_credentials": false,
-    "name": "Gringotts",
-    "process_status": {
-      "mfa": {
-        "credentials": [
+    most_recent_job_guid: 'JOB-241266ed-803a-4841-8a8a-0f37551e8f56',
+    needs_updated_credentials: false,
+    name: 'Gringotts',
+    process_status: {
+      mfa: {
+        credentials: [
           {
-            "nullify": null,
-            "guid": "CRD-241266ed-803a-4841-8a8a-0f37551e8f56",
-            "institution_guid": "INS-241266ed-803a-4841-8a8a-0f37551e8f56",
-            "label": "What city were you born in?",
-            "field_name": null,
-            "opt": null,
-            "type": 0,
-            "display_order": null,
-            "external_id": "UNIQUE_ID_FOR_THIS_CHALLENGE-f067cb4ab7",
-            "data_source_guid": null,
-            "field_type": 0,
-            "optional": null,
-            "editable": null,
-            "mfa": true,
-            "optional_mfa": null,
-            "aggregator_guid": "AGG-241266ed-803a-4841-8a8a-0f37551e8f56",
-            "question_field_type": null,
-            "answer_field_type": null,
-            "escaped": null,
-            "value_identifier": null,
-            "value_mask": null,
-            "size": null,
-            "max_length": null,
-            "created_at": "2023-03-30T22:45:43.000+00:00",
-            "updated_at": "2023-03-30T22:45:43.000+00:00",
-            "legacy_question_field_type": null,
-            "response_field_type": null,
-            "sequence": null,
-            "meta_data": null,
-            "display_name": null,
-            "status_code": 200,
-            "job_guid": "JOB-241266ed-803a-4841-8a8a-0f37551e8f56",
-            "options": []
+            nullify: null,
+            guid: 'CRD-241266ed-803a-4841-8a8a-0f37551e8f56',
+            institution_guid: 'INS-241266ed-803a-4841-8a8a-0f37551e8f56',
+            label: 'What city were you born in?',
+            field_name: null,
+            opt: null,
+            type: 0,
+            display_order: null,
+            external_id: 'UNIQUE_ID_FOR_THIS_CHALLENGE-f067cb4ab7',
+            data_source_guid: null,
+            field_type: 0,
+            optional: null,
+            editable: null,
+            mfa: true,
+            optional_mfa: null,
+            aggregator_guid: 'AGG-241266ed-803a-4841-8a8a-0f37551e8f56',
+            question_field_type: null,
+            answer_field_type: null,
+            escaped: null,
+            value_identifier: null,
+            value_mask: null,
+            size: null,
+            max_length: null,
+            created_at: '2023-03-30T22:45:43.000+00:00',
+            updated_at: '2023-03-30T22:45:43.000+00:00',
+            legacy_question_field_type: null,
+            response_field_type: null,
+            sequence: null,
+            meta_data: null,
+            display_name: null,
+            status_code: 200,
+            job_guid: 'JOB-241266ed-803a-4841-8a8a-0f37551e8f56',
+            options: []
           }
         ]
       }
     },
-    "revision": 11,
-    "user_guid": "USR-241266ed-803a-4841-8a8a-0f37551e8f56",
-    "oauth_window_uri": null,
-    "verification_is_enabled": true,
-    "tax_statement_is_enabled": false,
-    "successfully_aggregated_at": null
+    revision: 11,
+    user_guid: 'USR-241266ed-803a-4841-8a8a-0f37551e8f56',
+    oauth_window_uri: null,
+    verification_is_enabled: true,
+    tax_statement_is_enabled: false,
+    successfully_aggregated_at: null
   }
 }
