@@ -1,26 +1,26 @@
 module.exports = {
-  "member": {
-      "guid": "MBR-c48148e9-b06d-4cd0-9a9e-c927a53182fd",
-      "aggregated_at": "2023-05-06T00:06:20Z",
-      "has_processed_account_numbers": false,
-      "is_being_aggregated": true,
-      "successfully_aggregated_at": null,
-      "connection_status": "CHALLENGED",
-      "is_authenticated": false,
-      "challenges": [{
-              "field_name": "Request your single-use security code to be sent as a voice or text message.",
-              "guid": "CRD-9738fe85-b261-4b08-b148-7b8e9edda98a",
-              "label": "Request your single-use security code to be sent as a voice or text message.",
-              "type": "OPTIONS",
-              "options": [{
-                      "label": "Text me at (•••)•••-8582 ()",
-                      "value": "td_bank_text"
-                  }, {
-                      "label": "Call me at (•••)•••-8582 ()",
-                      "value": "td_bank_call"
-                  }
-              ]
-          }
+  member: {
+    guid: 'MBR-c48148e9-b06d-4cd0-9a9e-c927a53182fd',
+    aggregated_at: '2023-05-06T00:06:20Z',
+    has_processed_account_numbers: false,
+    is_being_aggregated: true,
+    successfully_aggregated_at: null,
+    connection_status: 'CHALLENGED',
+    is_authenticated: false,
+    challenges: [{
+      field_name: 'Request your single-use security code to be sent as a voice or text message.',
+      guid: 'CRD-9738fe85-b261-4b08-b148-7b8e9edda98a',
+      label: 'Request your single-use security code to be sent as a voice or text message.',
+      type: 'OPTIONS',
+      options: [{
+        label: 'Text me at (•••)•••-8582 ()',
+        value: 'td_bank_text'
+      }, {
+        label: 'Call me at (•••)•••-8582 ()',
+        value: 'td_bank_call'
+      }
       ]
+    }
+    ]
   }
 }
