@@ -1,29 +1,29 @@
 module.exports = {
-  "method": "POST",
-  "headers": {
-      "host": "webhook.sophtron-prod.com",
-      "x-request-id": "85c9e68f191856d485960f5bafcee9f5",
-      "x-real-ip": "34.215.116.35",
-      "x-forwarded-for": "34.215.116.35",
-      "x-forwarded-host": "webhook.sophtron-prod.com",
-      "x-forwarded-port": "80",
-      "x-forwarded-proto": "http",
-      "x-scheme": "http",
-      "content-length": "110",
-      "accept": "application/json, text/plain, */*",
-      "content-type": "application/json",
-      "x-finicity-signature": "7c9a654cc156e28802bd173e3c6ab2caebc89aa2827a4c69b9f32f092fef6be9",
-      "user-agent": "axios/0.25.0",
-      "x-b3-traceid": "0610bbd2cafbbfb010feb9b85175af16",
-      "x-b3-spanid": "c94c6d913e1f0663",
-      "x-b3-sampled": "1",
-      "traceparent": "00-0610bbd2cafbbfb010feb9b85175af16-c94c6d913e1f0663-01"
+  method: 'POST',
+  headers: {
+    host: 'webhook.sophtron-prod.com',
+    'x-request-id': '85c9e68f191856d485960f5bafcee9f5',
+    'x-real-ip': '34.215.116.35',
+    'x-forwarded-for': '34.215.116.35',
+    'x-forwarded-host': 'webhook.sophtron-prod.com',
+    'x-forwarded-port': '80',
+    'x-forwarded-proto': 'http',
+    'x-scheme': 'http',
+    'content-length': '110',
+    accept: 'application/json, text/plain, */*',
+    'content-type': 'application/json',
+    'x-finicity-signature': '7c9a654cc156e28802bd173e3c6ab2caebc89aa2827a4c69b9f32f092fef6be9',
+    'user-agent': 'axios/0.25.0',
+    'x-b3-traceid': '0610bbd2cafbbfb010feb9b85175af16',
+    'x-b3-spanid': 'c94c6d913e1f0663',
+    'x-b3-sampled': '1',
+    traceparent: '00-0610bbd2cafbbfb010feb9b85175af16-c94c6d913e1f0663-01'
   },
-  "path": "/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951",
-  "body": {
-      "customerId": "6030781868",
-      "eventType": "done",
-      "eventId": "1688829095490-47a556ac5d61cf3ad20e10c6",
-      "payload": {}
+  path: '/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951',
+  body: {
+    customerId: '6030781868',
+    eventType: 'done',
+    eventId: '1688829095490-47a556ac5d61cf3ad20e10c6',
+    payload: {}
   }
 }
