@@ -1,31 +1,31 @@
 module.exports = {
-  "method": "POST",
-  "headers": {
-      "host": "webhook.sophtron-prod.com",
-      "x-request-id": "9fc8577798e9c943c9f11575685f0612",
-      "x-real-ip": "34.215.116.35",
-      "x-forwarded-for": "34.215.116.35",
-      "x-forwarded-host": "webhook.sophtron-prod.com",
-      "x-forwarded-port": "80",
-      "x-forwarded-proto": "http",
-      "x-scheme": "http",
-      "content-length": "150",
-      "accept": "application/json, text/plain, */*",
-      "content-type": "application/json",
-      "x-finicity-signature": "f3d72d7e54188262d6a8cb14d0f198b9e5e5603f1b844d5cc6a3602d7a96c0e5",
-      "user-agent": "axios/0.25.0",
-      "x-b3-traceid": "a7dc288558c325bcea83b12fae6e6638",
-      "x-b3-spanid": "f8f0c08d1ece2b7c",
-      "x-b3-sampled": "1",
-      "traceparent": "00-a7dc288558c325bcea83b12fae6e6638-f8f0c08d1ece2b7c-01"
+  method: 'POST',
+  headers: {
+    host: 'webhook.sophtron-prod.com',
+    'x-request-id': '9fc8577798e9c943c9f11575685f0612',
+    'x-real-ip': '34.215.116.35',
+    'x-forwarded-for': '34.215.116.35',
+    'x-forwarded-host': 'webhook.sophtron-prod.com',
+    'x-forwarded-port': '80',
+    'x-forwarded-proto': 'http',
+    'x-scheme': 'http',
+    'content-length': '150',
+    accept: 'application/json, text/plain, */*',
+    'content-type': 'application/json',
+    'x-finicity-signature': 'f3d72d7e54188262d6a8cb14d0f198b9e5e5603f1b844d5cc6a3602d7a96c0e5',
+    'user-agent': 'axios/0.25.0',
+    'x-b3-traceid': 'a7dc288558c325bcea83b12fae6e6638',
+    'x-b3-spanid': 'f8f0c08d1ece2b7c',
+    'x-b3-sampled': '1',
+    traceparent: '00-a7dc288558c325bcea83b12fae6e6638-f8f0c08d1ece2b7c-01'
   },
-  "path": "/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951",
-  "body": {
-      "customerId": "6030781868",
-      "eventType": "institutionSupported",
-      "eventId": "1688829064021-3b2d9920bff89ae88d033e00",
-      "payload": {
-          "institutionId": "102176"
-      }
+  path: '/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951',
+  body: {
+    customerId: '6030781868',
+    eventType: 'institutionSupported',
+    eventId: '1688829064021-3b2d9920bff89ae88d033e00',
+    payload: {
+      institutionId: '102176'
+    }
   }
 }
