@@ -1,27 +1,27 @@
 module.exports = {
-  "method": "POST",
-  "headers": {
-      "host": "webhook.sophtron-prod.com",
-      "x-request-id": "5cba899887657958e578a3978d8ec576",
-      "x-real-ip": "34.215.116.35",
-      "x-forwarded-for": "34.215.116.35",
-      "x-forwarded-host": "webhook.sophtron-prod.com",
-      "x-forwarded-port": "80",
-      "x-forwarded-proto": "http",
-      "x-scheme": "http",
-      "content-length": "33",
-      "accept": "application/json, text/plain, */*",
-      "content-type": "application/json",
-      "x-finicity-signature": "3fb576b543beba215be394bfa53d9a2e896f7fdf44aac9509b70b4a4427f87f5",
-      "user-agent": "axios/0.25.0",
-      "x-b3-traceid": "66f62034218da84ec403cd06c324284d",
-      "x-b3-spanid": "d89844f3af4943e7",
-      "x-b3-sampled": "1",
-      "traceparent": "00-66f62034218da84ec403cd06c324284d-d89844f3af4943e7-01"
+  method: 'POST',
+  headers: {
+    host: 'webhook.sophtron-prod.com',
+    'x-request-id': '5cba899887657958e578a3978d8ec576',
+    'x-real-ip': '34.215.116.35',
+    'x-forwarded-for': '34.215.116.35',
+    'x-forwarded-host': 'webhook.sophtron-prod.com',
+    'x-forwarded-port': '80',
+    'x-forwarded-proto': 'http',
+    'x-scheme': 'http',
+    'content-length': '33',
+    accept: 'application/json, text/plain, */*',
+    'content-type': 'application/json',
+    'x-finicity-signature': '3fb576b543beba215be394bfa53d9a2e896f7fdf44aac9509b70b4a4427f87f5',
+    'user-agent': 'axios/0.25.0',
+    'x-b3-traceid': '66f62034218da84ec403cd06c324284d',
+    'x-b3-spanid': 'd89844f3af4943e7',
+    'x-b3-sampled': '1',
+    traceparent: '00-66f62034218da84ec403cd06c324284d-d89844f3af4943e7-01'
   },
-  "path": "/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951",
-  "body": {
-      "eventType": "ping",
-      "payload": {}
+  path: '/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951',
+  body: {
+    eventType: 'ping',
+    payload: {}
   }
 }
