@@ -1,31 +1,31 @@
 module.exports = {
-  "method": "POST",
-  "headers": {
-      "host": "webhook.sophtron-prod.com",
-      "x-request-id": "d4c7c4272203b9cba75b1c29d27b991a",
-      "x-real-ip": "34.215.116.35",
-      "x-forwarded-for": "34.215.116.35",
-      "x-forwarded-host": "webhook.sophtron-prod.com",
-      "x-forwarded-port": "80",
-      "x-forwarded-proto": "http",
-      "x-scheme": "http",
-      "content-length": "140",
-      "accept": "application/json, text/plain, */*",
-      "content-type": "application/json",
-      "x-finicity-signature": "d43e9dd5b41ae3b17b892a39825c96f5f97554ec686140150589077a0bf24276",
-      "user-agent": "axios/0.25.0",
-      "x-b3-traceid": "afd7418bc7f0ac40a2214f947870b23b",
-      "x-b3-spanid": "a022d23582645673",
-      "x-b3-sampled": "1",
-      "traceparent": "00-afd7418bc7f0ac40a2214f947870b23b-a022d23582645673-01"
+  method: 'POST',
+  headers: {
+    host: 'webhook.sophtron-prod.com',
+    'x-request-id': 'd4c7c4272203b9cba75b1c29d27b991a',
+    'x-real-ip': '34.215.116.35',
+    'x-forwarded-for': '34.215.116.35',
+    'x-forwarded-host': 'webhook.sophtron-prod.com',
+    'x-forwarded-port': '80',
+    'x-forwarded-proto': 'http',
+    'x-scheme': 'http',
+    'content-length': '140',
+    accept: 'application/json, text/plain, */*',
+    'content-type': 'application/json',
+    'x-finicity-signature': 'd43e9dd5b41ae3b17b892a39825c96f5f97554ec686140150589077a0bf24276',
+    'user-agent': 'axios/0.25.0',
+    'x-b3-traceid': 'afd7418bc7f0ac40a2214f947870b23b',
+    'x-b3-spanid': 'a022d23582645673',
+    'x-b3-sampled': '1',
+    traceparent: '00-afd7418bc7f0ac40a2214f947870b23b-a022d23582645673-01'
   },
-  "path": "/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951",
-  "body": {
-      "customerId": "6030781868",
-      "eventType": "processing",
-      "eventId": "1688829079076-762bbd85f01689ccfa615f7f",
-      "payload": {
-          "institutionId": "102176"
-      }
+  path: '/webhook/finicity_sandbox/f0105b8f-a7a4-4665-98e2-09e116804951',
+  body: {
+    customerId: '6030781868',
+    eventType: 'processing',
+    eventId: '1688829079076-762bbd85f01689ccfa615f7f',
+    payload: {
+      institutionId: '102176'
+    }
   }
 }
