@@ -1,22 +1,22 @@
 module.exports = {
-  "credentials": [{
-          "display_order": 1,
-          "field_name": "LOGIN",
-          "field_type": 3,
-          "guid": "CRD-9f61fb4c-912c-bd1e-b175-ccc7f0275cc1",
-          "label": "Username",
-          "meta_data": null,
-          "optional": false,
-          "options": null
-      }, {
-          "display_order": 2,
-          "field_name": "PASSWORD",
-          "field_type": 1,
-          "guid": "CRD-e3d7ea81-aac7-05e9-fbdd-4b493c6e474d",
-          "label": "Password",
-          "meta_data": null,
-          "optional": false,
-          "options": null
-      }
+  credentials: [{
+    display_order: 1,
+    field_name: 'LOGIN',
+    field_type: 3,
+    guid: 'CRD-9f61fb4c-912c-bd1e-b175-ccc7f0275cc1',
+    label: 'Username',
+    meta_data: null,
+    optional: false,
+    options: null
+  }, {
+    display_order: 2,
+    field_name: 'PASSWORD',
+    field_type: 1,
+    guid: 'CRD-e3d7ea81-aac7-05e9-fbdd-4b493c6e474d',
+    label: 'Password',
+    meta_data: null,
+    optional: false,
+    options: null
+  }
   ]
 }
