@@ -1,20 +1,20 @@
 module.exports = {
-  "members": [
+  members: [
     {
-        "guid": "MBR-241266ed-803a-4841-8a8a-0f37551e8f56",
-        "id": null,
-        "user_guid": "USR-241266ed-803a-4841-8a8a-0f37551e8f56",
-        "user_id": "241266ed-803a-4841-8a8a-0f37551e8f56",
-        "aggregated_at": "2023-03-01T16:46:00Z",
-        "background_aggregation_is_disabled": false,
-        "institution_code": "gringotts",
-        "is_being_aggregated": false,
-        "is_managed_by_user": true,
-        "is_oauth": false,
-        "metadata": null,
-        "name": "Gringotts",
-        "successfully_aggregated_at": "2023-03-01T16:46:06Z",
-        "connection_status": "CONNECTED"
+      guid: 'MBR-241266ed-803a-4841-8a8a-0f37551e8f56',
+      id: null,
+      user_guid: 'USR-241266ed-803a-4841-8a8a-0f37551e8f56',
+      user_id: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      aggregated_at: '2023-03-01T16:46:00Z',
+      background_aggregation_is_disabled: false,
+      institution_code: 'gringotts',
+      is_being_aggregated: false,
+      is_managed_by_user: true,
+      is_oauth: false,
+      metadata: null,
+      name: 'Gringotts',
+      successfully_aggregated_at: '2023-03-01T16:46:06Z',
+      connection_status: 'CONNECTED'
     }
   ]
 }
