@@ -1,74 +1,74 @@
 module.exports = {
-  "user_features": [
+  user_features: [
     {
-      "feature_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "feature_name": "SHOW_NEW_MANUAL_ACCOUNTS_CONNECT",
-      "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "is_enabled": true,
-      "should_notify_user": false,
-      "user_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "created_at": "2022-09-07T19:20:11.000Z",
-      "updated_at": "2022-12-13T23:29:34.000Z"
+      feature_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      feature_name: 'SHOW_NEW_MANUAL_ACCOUNTS_CONNECT',
+      guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      is_enabled: true,
+      should_notify_user: false,
+      user_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      created_at: '2022-09-07T19:20:11.000Z',
+      updated_at: '2022-12-13T23:29:34.000Z'
     },
     {
-      "feature_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "feature_name": "CONNECT_HIDE_LIGHT_DISCLOSURE",
-      "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "is_enabled": true,
-      "should_notify_user": true,
-      "user_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "created_at": "2022-10-05T18:01:51.000Z",
-      "updated_at": "2022-10-05T18:01:51.000Z"
+      feature_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      feature_name: 'CONNECT_HIDE_LIGHT_DISCLOSURE',
+      guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      is_enabled: true,
+      should_notify_user: true,
+      user_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      created_at: '2022-10-05T18:01:51.000Z',
+      updated_at: '2022-10-05T18:01:51.000Z'
     },
     {
-      "feature_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "feature_name": "CONNECT_OAUTH_WINDOW_POPUP",
-      "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "is_enabled": true,
-      "should_notify_user": true,
-      "user_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "created_at": "2022-10-10T20:41:23.000Z",
-      "updated_at": "2022-10-10T20:41:23.000Z"
+      feature_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      feature_name: 'CONNECT_OAUTH_WINDOW_POPUP',
+      guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      is_enabled: true,
+      should_notify_user: true,
+      user_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      created_at: '2022-10-10T20:41:23.000Z',
+      updated_at: '2022-10-10T20:41:23.000Z'
     },
     {
-      "feature_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "feature_name": "CONNECT_GO_TO_INSTITUTION",
-      "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "is_enabled": true,
-      "should_notify_user": false,
-      "user_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "created_at": "2022-10-19T16:39:51.000Z",
-      "updated_at": "2023-01-04T18:43:37.000Z"
+      feature_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      feature_name: 'CONNECT_GO_TO_INSTITUTION',
+      guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      is_enabled: true,
+      should_notify_user: false,
+      user_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      created_at: '2022-10-19T16:39:51.000Z',
+      updated_at: '2023-01-04T18:43:37.000Z'
     },
     {
-      "feature_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "feature_name": "CONNECT_SECURE_SIGN_IN_NEW",
-      "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "is_enabled": true,
-      "should_notify_user": true,
-      "user_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "created_at": "2022-10-25T20:46:51.000Z",
-      "updated_at": "2023-01-03T17:32:45.000Z"
+      feature_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      feature_name: 'CONNECT_SECURE_SIGN_IN_NEW',
+      guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      is_enabled: true,
+      should_notify_user: true,
+      user_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      created_at: '2022-10-25T20:46:51.000Z',
+      updated_at: '2023-01-03T17:32:45.000Z'
     },
     {
-      "feature_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "feature_name": "CONNECT_SURVEY_PROMPT",
-      "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "is_enabled": true,
-      "should_notify_user": false,
-      "user_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "created_at": "2022-12-09T20:56:17.000Z",
-      "updated_at": "2022-12-09T20:56:17.000Z"
+      feature_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      feature_name: 'CONNECT_SURVEY_PROMPT',
+      guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      is_enabled: true,
+      should_notify_user: false,
+      user_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      created_at: '2022-12-09T20:56:17.000Z',
+      updated_at: '2022-12-09T20:56:17.000Z'
     },
     {
-      "feature_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "feature_name": "SHOW_FINSTRONG_WIDGET",
-      "guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "is_enabled": true,
-      "should_notify_user": false,
-      "user_guid": "241266ed-803a-4841-8a8a-0f37551e8f56",
-      "created_at": "2023-02-10T19:22:59.000Z",
-      "updated_at": "2023-02-10T19:22:59.000Z"
+      feature_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      feature_name: 'SHOW_FINSTRONG_WIDGET',
+      guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      is_enabled: true,
+      should_notify_user: false,
+      user_guid: '241266ed-803a-4841-8a8a-0f37551e8f56',
+      created_at: '2023-02-10T19:22:59.000Z',
+      updated_at: '2023-02-10T19:22:59.000Z'
     }
   ]
 }
