@@ -1,20 +1,20 @@
 import {
-  Institution,
-} from '../../shared/contract';
-import * as config from '../config';
-import * as logger from '../infra/logger';
+  type Institution
+} from '../../shared/contract'
+import * as config from '../config'
+import * as logger from '../infra/logger'
 
-import { ProviderApiBase } from '../providers';
+import { ProviderApiBase } from '../providers'
 
 export class VcsService extends ProviderApiBase {
-  constructor(req: any){
+  constructor (req: any) {
     super(req)
   }
 
-  async selectInstitution(
-    institution: Institution,
+  async selectInstitution (
+    institution: Institution
   ) {
-    logger.debug(`Selecting institution ${institution.id}`);
+    logger.debug(`Selecting institution ${institution.id}`)
     // if (institution.provider) {
     //   context.provider = institution.provider;
     // }
@@ -22,57 +22,57 @@ export class VcsService extends ProviderApiBase {
     //   context.provider = 'sophtron';
     // }
     if (institution.id) {
-      institution = await this.getInstitution(institution.id);
-      const credentials = await this.getInstitutionCredentials(institution.id);
-      return { institution, credentials };
+      institution = await this.getInstitution(institution.id)
+      const credentials = await this.getInstitutionCredentials(institution.id)
+      return { institution, credentials }
     }
-    return { error: 'invalid institution selected' };
+    return { error: 'invalid institution selected' }
   }
 
-  async login(
+  async login (
     institution_id: string,
     connection_id: string | null,
-    credentials: Array<Credential>
+    credentials: Credential[]
   ) {
-    institution_id = this.context.institution_id || institution_id;
-    connection_id = this.context.connection_id || connection_id;
+    institution_id = this.context.institution_id || institution_id
+    connection_id = this.context.connection_id || connection_id
     if (connection_id) {
       const res = await this.updateConnection(
         {
           id: connection_id,
-          credentials,
+          credentials
         }
-      );
-      return res;
+      )
+      return res
     }
     if (institution_id) {
       const res = await this.createConnection(
         {
           institution_id,
           credentials,
-          initial_job_type: this.context.job_type,
+          initial_job_type: this.context.job_type
         }
-      );
+      )
       if (res) {
-        return res;
+        return res
       }
       logger.error(
         `failed creating connection with instituionId : ${institution_id}`
-      );
+      )
       return {
-        error: `failed creating connection with instituionId : ${institution_id}`,
-      };
+        error: `failed creating connection with instituionId : ${institution_id}`
+      }
     }
-    return { error: 'Unable to find instituion, invalid parameters provided' };
+    return { error: 'Unable to find instituion, invalid parameters provided' }
   }
 
-  async mfa(id: string) {
-    const res = await this.getConnectionStatus(id);
+  async mfa (id: string) {
+    const res = await this.getConnectionStatus(id)
     if (!res) {
-      logger.warning(`Mfa failed to find connection with Id: ${id}`);
-      return { error: 'Failed to find job' };
+      logger.warning(`Mfa failed to find connection with Id: ${id}`)
+      return { error: 'Failed to find job' }
     }
-    res.provider = this.context.provider;
-    return res;
+    res.provider = this.context.provider
+    return res
   }
 }
