@@ -1,9 +1,9 @@
-const config = require('../../config');
-const real = require('./real');
-const mocked = require('./mock');
+const config = require('../../config')
+const real = require('./real')
+const mocked = require('./mock')
 
-const mock = config.Env === 'mocked';
+const mock = config.Env === 'mocked'
 
-const http = mock ? mocked : real;
+const http = mock ? mocked : real
 
-module.exports = http;
+module.exports = http
