@@ -1,58 +1,58 @@
-const config = require('../../config');
-const logger = require('../../infra/logger');
-const http = require('../../infra/http');
-const SophtronBaseClient = require('./base');
+const config = require('../../config')
+const logger = require('../../infra/logger')
+const http = require('../../infra/http')
+const SophtronBaseClient = require('./base')
 
-module.exports = class SophtronV2Client extends SophtronBaseClient{
-  constructor(apiConfig){
-    super(apiConfig);
+module.exports = class SophtronV2Client extends SophtronBaseClient {
+  constructor (apiConfig) {
+    super(apiConfig)
   }
 
-  getCustomer(customerId){
-    return this.get(`/v2/customers/${customerId}`)
+  async getCustomer (customerId) {
+    return await this.get(`/v2/customers/${customerId}`)
   }
 
-  async getCustomerByUniqueName(unique_name){
+  async getCustomerByUniqueName (unique_name) {
     const arr = await this.get(`/v2/customers?uniqueID=${unique_name}`)
-    return arr?.[0];
+    return arr?.[0]
   }
 
-  createCustomer(unique_name){
-    return this.post(`/v2/customers`,{
+  async createCustomer (unique_name) {
+    return await this.post('/v2/customers', {
       UniqueID: unique_name,
       Source: `Universal_Widget_${config.HostUrl}`,
       Name: 'UniversalWidget_Customer'
     })
   }
 
-  getMember(customerId, memberId){
-    return this.get(`/v2/customers/${customerId}/members/${memberId}`)
+  async getMember (customerId, memberId) {
+    return await this.get(`/v2/customers/${customerId}/members/${memberId}`)
   }
 
-  createMember(customerId, job_type, username, password, institution_id){
-    return this.post(`/v2/customers/${customerId}/members/${job_type}`, {
+  async createMember (customerId, job_type, username, password, institution_id) {
+    return await this.post(`/v2/customers/${customerId}/members/${job_type}`, {
       UserName: username,
       Password: password,
       InstitutionID: institution_id
     })
   }
 
-  updateMember(customerId, memberId, job_type, username, password){
-    return this.put(`/v2/customers/${customerId}/members/${memberId}/${job_type}`, {
+  async updateMember (customerId, memberId, job_type, username, password) {
+    return await this.put(`/v2/customers/${customerId}/members/${memberId}/${job_type}`, {
       UserName: username,
       Password: password
     })
   }
 
-  refreshMember(customerId, memberId, job_type){
-    return this.post(`/v2/customers/${customerId}/members/${memberId}/${job_type}`)
+  async refreshMember (customerId, memberId, job_type) {
+    return await this.post(`/v2/customers/${customerId}/members/${memberId}/${job_type}`)
   }
 
-  deleteMember(customerId, memberId){
-    return this.del(`/v2/customers/${customerId}/members/${memberId}`)
+  async deleteMember (customerId, memberId) {
+    return await this.del(`/v2/customers/${customerId}/members/${memberId}`)
   }
 
-  getJobInfo(jobId){
-    return this.get(`/v2/job/${jobId}`);
+  async getJobInfo (jobId) {
+    return await this.get(`/v2/job/${jobId}`)
   }
-};
+}
