@@ -1,29 +1,28 @@
-
 export interface Credential {
   guid: string //
   value?: string
   nullify?: null
-  institution_guid?: string //INS-asdf
-  label?: string //What city were you born in?
-  field_name?: string //null
-  opt?: string //null
+  institution_guid?: string // INS-asdf
+  label?: string // What city were you born in?
+  field_name?: string // null
+  opt?: string // null
   type?: number // 0
-  display_order?: any //null
+  display_order?: any // null
   external_id?: string // UNIQUE_ID_FOR_THIS_CHALLENGE-6a8d4e0395
   data_source_guid?: string // null
   field_type?: number // 0
-  optional?: any //null
+  optional?: any // null
   editable?: any // null
   mfa?: boolean // true
   optional_mfa?: any // null
   aggregator_guid?: string // AGG-asdf
-  question_field_type?: any //null
+  question_field_type?: any // null
   answer_field_type?: any // null
   escaped?: any // null
   value_identifier?: any // null
-  value_mask?: any //null
+  value_mask?: any // null
   size?: any // null
-  max_length?: any //null
+  max_length?: any // null
   created_at?: string // 2023-03-02T22:08:47.000+00:00
   updated_at?: string // 2023-03-02T22:08:47.000+00:00
   legacy_question_field_type?: any // null
@@ -33,14 +32,14 @@ export interface Credential {
   display_name?: string // null
   status_code?: number // 200
   job_guid?: string // JOB-asdf
-  options?: any //Array // []
+  options?: any // Array // []
 }
 
 export interface Member {
-  //aggregation_status: number //
-  institution_guid: string //INS-asdf
+  // aggregation_status: number //
+  institution_guid: string // INS-asdf
   institution_name?: string //
-  institution_url?: string //https://gringotts.qa.internal.mx/
+  institution_url?: string // https://gringotts.qa.internal.mx/
   instructional_data?: object
   // {
   //   title?: Please do all these things <a href=\https://google.com\ id=\instructional_text\>My Test Link</a>
@@ -52,39 +51,39 @@ export interface Member {
   //   ]
   // }
   is_manual?: boolean //
-  last_job_guid?: string //JOB-asdf
+  last_job_guid?: string // JOB-asdf
   last_job_status?: string //
   last_update_time?: string //
 
-  most_recent_job_guid?: string //JOB-asdf
+  most_recent_job_guid?: string // JOB-asdf
   needs_updated_credentials?: boolean
-  revision?: number //13
+  revision?: number // 13
   verification_is_enabled?: boolean
   tax_statement_is_enabled?: boolean
-  aggregated_at?: string | null;
-  background_aggregation_is_disabled?: boolean;
-  connection_status?: string | null | number; //
-  guid: string | null; //
-  id?: string | null;
-  institution_code?: string | null;
-  is_being_aggregated?: boolean | null; //
-  is_managed_by_user?: boolean | null; //
-  is_oauth?: boolean | null; //
-  metadata?: string | null; //
-  name?: string | null; //
-  oauth_window_uri?: string | null; //
-  successfully_aggregated_at?: string | null; //
-  user_guid?: string | null; //
-  user_id?: string | null;
+  aggregated_at?: string | null
+  background_aggregation_is_disabled?: boolean
+  connection_status?: string | null | number //
+  guid: string | null //
+  id?: string | null
+  institution_code?: string | null
+  is_being_aggregated?: boolean | null //
+  is_managed_by_user?: boolean | null //
+  is_oauth?: boolean | null //
+  metadata?: string | null //
+  name?: string | null //
+  oauth_window_uri?: string | null //
+  successfully_aggregated_at?: string | null //
+  user_guid?: string | null //
+  user_id?: string | null
   mfa?: {
-    credentials: Array<Credential>
+    credentials: Credential[]
   }
   process_status?: {
-    credentials: Array<Credential>
+    credentials: Credential[]
   }
-  skip_aggregation?: boolean | null;
+  skip_aggregation?: boolean | null
   credentials?: Credential[]
 }
-export interface MemberResponse{
-  member: Member;
+export interface MemberResponse {
+  member: Member
 }
